Add disqus_identifier variable as the relative path
If people visit your website from facebook it may add some URL
querystring that makes Disqus think it's a different webpage,
consequently creating a duplicate discussion for the same article.
This variable prevents that by specifically declaring
window.location.pathname as the discussion identifier, which should be the
same regardless of querystrings.