Removed superfluous tags from sidebar image section
The sidebar image is positioned too close to the "show/hide sidebar control". I put some tags there that introduced an empty space, to push the sidebar image downwards. However, the "right way" to manage that space is through CSS. So in this commit, I removed said tags to make the code cleaner.