author's name gets link only if AUTHORS_SAVE_AS is set.
There seems to be no point in linking to the author's index page listing all
articles of that author if there is only one author. If the pelican
configuration states that no overview page of all authors should be generated
we can assume only one author exists and thus not link the author's name in the
article footer.