Load scripts before plugins that might use them.
Body elements are loaded in the order that they are seen; by loading
at the end of the body, they are not available for content that needs them,
like the github plugin in the sidebar. For the github plugin, this results
in a '$ is not defined' error and the repo list is never updated. Moving
the scripts to <head> ensures the scripts are loaded before dependent
content.